seo-augmenter/backend/requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
duckduckgo-search
beautifulsoup4
requests
html2text
